+++ homeassistant/components/jvc_projector/const.py
