+++ homeassistant/components/jvc_projector/coordinator.py
