+++ homeassistant/components/jvc_projector/entity.py
