+++ homeassistant/components/jvc_projector/manifest.json
@@ -1,8 +1,8 @@
 {
-  "domain": "jvcprojector",
+  "domain": "jvc_projector",
   "name": "JVC Projector",
   "config_flow": true,
-  "documentation": "https://www.home-assistant.io/integrations/jvcprojector",
+  "documentation": "https://www.home-assistant.io/integrations/jvc_projector",
   "requirements": ["pyjvcprojector==1.0.0"],
   "codeowners": ["@SteveEasley"],
   "iot_class": "local_polling"
