+++ homeassistant/components/jvc_projector/remote.py
