+++ homeassistant/components/jvc_projector/sensor.py
