+++ homeassistant/components/jvc_projector/strings.json
