+++ homeassistant/components/jvc_projector/translations/bg.json
