+++ homeassistant/components/jvc_projector/translations/ca.json
