+++ homeassistant/components/jvc_projector/translations/cs.json
