+++ homeassistant/components/jvc_projector/translations/de.json
