+++ homeassistant/components/jvc_projector/translations/el.json
