+++ homeassistant/components/jvc_projector/translations/en.json
