+++ homeassistant/components/jvc_projector/translations/es.json
