+++ homeassistant/components/jvc_projector/translations/et.json
