+++ homeassistant/components/jvc_projector/translations/fr.json
