+++ homeassistant/components/jvc_projector/translations/he.json
