+++ homeassistant/components/jvc_projector/translations/hu.json
