+++ homeassistant/components/jvc_projector/translations/id.json
