+++ homeassistant/components/jvc_projector/translations/it.json
