+++ homeassistant/components/jvc_projector/translations/ja.json
