+++ homeassistant/components/jvc_projector/translations/ko.json
