+++ homeassistant/components/jvc_projector/translations/nb.json
