+++ homeassistant/components/jvc_projector/translations/nl.json
