+++ homeassistant/components/jvc_projector/translations/no.json
