+++ homeassistant/components/jvc_projector/translations/pl.json
