+++ homeassistant/components/jvc_projector/translations/pt-BR.json
