+++ homeassistant/components/jvc_projector/translations/pt.json
