+++ homeassistant/components/jvc_projector/translations/ru.json
