+++ homeassistant/components/jvc_projector/translations/sk.json
