+++ homeassistant/components/jvc_projector/translations/sv.json
