+++ homeassistant/components/jvc_projector/translations/tr.json
