+++ homeassistant/components/jvc_projector/translations/zh-Hant.json
