+++ tests/components/jvc_projector/__init__.py
