+++ tests/components/jvc_projector/test_coordinator.py
