+++ tests/components/jvc_projector/test_remote.py
